+++ src/video_core/renderer_opengl/gl_rasterizer.h
@@ -32,9 +32,9 @@ public:
     void DrawTriangles() override;
     void NotifyMaxwellRegisterChanged(u32 id) override;
     void FlushAll() override;
-    void FlushRegion(VAddr addr, u32 size) override;
-    void InvalidateRegion(VAddr addr, u32 size) override;
-    void FlushAndInvalidateRegion(VAddr addr, u32 size) override;
+    void FlushRegion(VAddr addr, u64 size) override;
+    void InvalidateRegion(VAddr addr, u64 size) override;
+    void FlushAndInvalidateRegion(VAddr addr, u64 size) override;
     bool AccelerateDisplayTransfer(const void* config) override;
     bool AccelerateTextureCopy(const void* config) override;
     bool AccelerateFill(const void* config) override;
